Code Reorganisation
Took inspiration from the other pipelines
