+++ lib/index.js
@@ -1,51 +1,80 @@
-let buildBundler = require("./build-bundler");
-let { abort, repr } = require("faucet-pipeline-core/lib/util");
+let { abort } = require("faucet-pipeline-core/lib/util");
+let path = require("path");
+let makeAutoprefixer = require("./make-autoprefixer");
+let makeSassRenderer = require("./make-sass-renderer");
 
-module.exports = (pluginConfig, assetManager, options) => {
-	let { browsers, compact, sourcemaps } = options;
-	let config = buildConfig(pluginConfig, assetManager, {
-		browsers, compact, sourcemaps
-	});
-	return buildBundleAll(config);
+module.exports = (config, assetManager, { browsers, compact, sourcemaps } = {}) => {
+	let bundlers = config.map(bundleConfig =>
+		makeBundler(bundleConfig, assetManager, { browsers, compact, sourcemaps }));
+
+	return filepaths => Promise.all(bundlers.map(bundle => bundle(filepaths)));
 };
 
-function buildConfig(pluginConfig, assetManager, { browsers, compact, sourcemaps }) {
-	return pluginConfig.map(({ source, target, browserslist, fingerprint }) => {
-		if(!source || !target) {
-			let setting = source ? "target" : "source";
-			abort(`ERROR: missing ${repr(setting, false)} configuration in Sass bundle`);
+function makeBundler(config, assetManager, { browsers, compact, sourcemaps } = {}) {
+	let { browserslist, fingerprint } = config;
+
+	if(!config.source || !config.target) {
+		abort("ERROR: Sass configuration requires both target and source");
+	}
+
+	let source = assetManager.resolvePath(config.source);
+	let target = assetManager.resolvePath(config.target, { enforceRelative: true });
+
+	if(browserslist === false) {
+		browsers = null;
+	} else if(browserslist) {
+		browsers = browsers[browserslist];
+	} else {
+		browsers = browsers.defaults;
+	}
+
+	let renderSass = makeSassRenderer(source, assetManager, sourcemaps, compact);
+	let autoprefix = makeAutoprefixer(source, assetManager, sourcemaps, browsers);
+
+	let previouslyIncludedFiles;
+
+	return filepaths => {
+		// If this is the first run or the changed file is one of the
+		// previously included ones, run the compiler
+		if(previouslyIncluded(filepaths, previouslyIncludedFiles)) {
+			renderSass().
+				then(autoprefix).
+				then(result => {
+					previouslyIncludedFiles = result.stats.includedFiles.
+						map(filepath => path.normalize(filepath));
+
+					let options = {};
+					if(fingerprint !== undefined) {
+						options.fingerprint = fingerprint;
+					}
+					assetManager.writeFile(target, result.css, options);
+				}).
+				catch(error => {
+					let options = { error };
+					if(fingerprint !== undefined) {
+						options.fingerprint = fingerprint;
+					}
+					assetManager.writeFile(target, errorOutput(error.message),
+							options);
+				});
 		}
-
-		let inputFileName = assetManager.resolvePath(source);
-		target = assetManager.resolvePath(target, { enforceRelative: true });
-
-		let outputStyle = compact ? "compact" : "nested";
-		let includePaths = [ assetManager.packagesDir ];
-
-		let selectedBrowsers;
-		if(browserslist === false) {
-			selectedBrowsers = null;
-		} else if(browserslist) {
-			selectedBrowsers = browsers[browserslist];
-		} else {
-			selectedBrowsers = browsers.defaults;
-		}
-
-		return {
-			inputFileName,
-			includePaths,
-			browsers: selectedBrowsers,
-			outputStyle,
-			assetManager,
-			fingerprint,
-			sourcemaps,
-			target
-		};
-	});
+	};
 }
 
-function buildBundleAll(config) {
-	let bundlers = config.map(bundleConfig => buildBundler(bundleConfig));
+function previouslyIncluded(filepaths, previouslyIncludedFiles) {
+	return previouslyIncludedFiles === undefined ||
+		filepaths.some(filepath => {
+			return previouslyIncludedFiles.
+				find(candidate => candidate === path.normalize(filepath));
+		});
+}
 
-	return files => bundlers.forEach(bundler => bundler(files));
+function errorOutput(message) {
+	return `body:before {
+	content: "\\26a0  CSS Error: ${message.replace(/"/g, "'")}";
+	font-weight: bold;
+	display: block;
+	border: 5px solid red;
+	padding: 5px;
+}\n`;
 }
